user_service/requirements.txt
# user_service/requirements.txt
# Web framework
fastapi==0.104.1
starlette==0.27.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
SQLAlchemy==2.0.20
alembic==1.16.5
psycopg2-binary==2.9.7

# HTTP clients
requests==2.31.0
httpx==0.27.2

# Security/Auth
python-jose==3.5.0
bcrypt==4.3.0
PyJWT==2.10.1
passlib[bcrypt]==1.7.4

# Pydantic
pydantic==2.11.9
pydantic-settings==2.8.1
email-validator==2.1.0

# Environment
python-dotenv==1.0.0

# Utilities
python-dateutil==2.9.0.post0
tenacity==8.5.0
